Pass only boolean 'testDependency' argument to IdeDependenciesVisitor
