+++ subprojects/ide/src/main/java/org/gradle/plugins/ide/internal/resolver/IdeDependencyVisitor.java
@@ -16,7 +16,6 @@
 
 package org.gradle.plugins.ide.internal.resolver;
 
-import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.artifacts.result.ResolvedArtifactResult;
 import org.gradle.api.artifacts.result.UnresolvedDependencyResult;
 
@@ -53,12 +52,12 @@ public interface IdeDependencyVisitor {
      * The component identifier is guaranteed to be a {@link org.gradle.api.artifacts.component.ModuleComponentIdentifier}.
      * The source and javadoc locations maybe be empty, but never null.
      */
-    void visitModuleDependency(ResolvedArtifactResult artifact, Set<ResolvedArtifactResult> sources, Set<ResolvedArtifactResult> javaDoc, Set<Configuration> configurations);
+    void visitModuleDependency(ResolvedArtifactResult artifact, Set<ResolvedArtifactResult> sources, Set<ResolvedArtifactResult> javaDoc, boolean testDependency);
 
     /**
      * The dependency points neither to a project, nor an external module, so this method should treat it as an opaque file.
      */
-    void visitFileDependency(ResolvedArtifactResult artifact, Set<Configuration> configurations);
+    void visitFileDependency(ResolvedArtifactResult artifact, boolean testDependency);
 
     /**
      * There was an unresolved dependency in the result.
